Move test about parsing error to fixture about parsing errors
